PR: Change the path of QWatson default config directory (#46)
* Change the path of the default config directory

* Add a note in the Readme

* Make method name more explicit

* Clean imports

* Update README.md

* Add a new CollapsableScrollArea class

A QScrollArea in which the minimumSizeHint and sizeHint methods have
beed overriden in order to force the sie hint to respect the
minimumHeight set by the user.

* Rework args and layout of InfoBox

* if __name__ == '__main__':

* Update CloseDialog to the InfoBox changes

* Update DateTimeInputDialog to the InfoBox changes

* Do not use a strong ref to client.frames

If frames is reseted in the client for whatever reason, this will break
the link between the model and the data.

* New utils func to reset the client internal vars

* Add a method to remove widget from the stack

* Add more icon sizes

* Minor tweak to the layout

* New ImportDialog and QWatsonImportMixin

* Add the new dialog to the main window

* Rework the layout of the dialogs

* Create a base class for the dialogs

* Use BaseDialog in CloseDialog and upd tests

* Use BaseDialog in ImportDialog

* Clean imports and remove duplicate default btn

* Use BaseDialog in DateTimeInputDialog and upd tests

* Add tests and fix bugs

* Upd from master and upd the new tests

* Update README.md

* Assert the overview table is updated correctly

+++ qwatson/utils/fileio.py
@@ -10,6 +10,7 @@
 # ---- Stantard imports
 
 import os
+import os.path as osp
 from shutil import rmtree
 
 
@@ -25,9 +26,10 @@ def delete_file_safely(filename):
 
 def delete_folder_recursively(dirpath):
     """Try to delete all files and sub-folders below the given dirpath."""
-    for filename in os.listdir(dirpath):
-        filepath = os.path.join(dirpath, filename)
-        try:
-            rmtree(filepath)
-        except OSError:
-            os.remove(filepath)
+    if osp.exists(dirpath):
+        for filename in os.listdir(dirpath):
+            filepath = os.path.join(dirpath, filename)
+            try:
+                rmtree(filepath)
+            except OSError:
+                delete_file_safely(filepath)
